refactor: 코드 수정 반영
